hardware/requirements.txt
Adafruit-Blinka==9.0.4
adafruit-circuitpython-busdevice==5.2.16
adafruit-circuitpython-connectionmanager==3.1.7
adafruit-circuitpython-pn532==2.4.8
adafruit-circuitpython-requests==4.1.16
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.88.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
anyio==4.13.0
binho-host-adapter==0.1.6
board==1.0
certifi==2026.2.25
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.30.0
paho-mqtt==2.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pyftdi==0.57.1
pyserial==3.5
python-dotenv==1.2.2
pyusb==1.3.1
smbus2==0.6.0
sniffio==1.3.1
SpeechRecognition==3.15.2
sysv_ipc==1.2.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
VL53L1X==0.0.5
